fix MdsDataProvider FrameData
- time = GetFloatArray("DIM_OF( _jScope_img, 0 )"); did not work on 0Q
- changed all frame times to type double

+++ javascope/jScope/FrameData.java
@@ -60,7 +60,7 @@ interface FrameData
      * @return The time array for the frame sequence.
      * @exception java.io.IOException
      */
-    public float[] GetFrameTimes() throws Exception;
+    public double[] GetFrameTimes() throws Exception;
 
     /**
      * Return the frame at the given position.
